rename Trait component
